Add alias and rename the action log plugin base (#25445)
